requirements.txt
# Core dependencies for novelty assessment system

# LangChain ecosystem
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.5

# LLM and API clients
openai>=1.0.0
litellm>=1.0.0

# Data processing and validation
pydantic>=2.0.0
numpy>=1.24.0
requests>=2.28.0

# Text processing and similarity
rapidfuzz>=3.0.0
sentence-transformers>=2.2.0

# Scientific computing
scikit-learn>=1.3.0

# XML processing (for GROBID TEI files)
lxml>=4.9.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.0.0

# Paper retrieval (ACL Anthology)
acl-anthology>=0.2.0

mineru[core]

